+++ src/components/home/SectionInfo.tsx
