+++ src/components/home/SectionMission.tsx
